Migrate 'is' test
